Crate dmar

Definitions for the DMAR, the Direct Memory Access (DMA) Remapping ACPI table.

The structures defined herein are based on Section 8 of Intel’s VT Directed I/O Spec: https://software.intel.com/content/www/us/en/develop/download/intel-virtualization-technology-for-directed-io-architecture-specification.html

Structs

DeviceScopePath

Dmar
A wrapper around the DMAR ACPI table ([DmarReporting]), which contains details about IOMMU configuration.

DmarAndd
ANDD: DMAR ACPI Name-space Device Declaration Structure.

DmarAtsr
ATSR: DMAR Root Port ATS (Address Translation Services) Capability Reporting Structure.

DmarDeviceScope
DMAR Device Scope Structure.

DmarDrhd
DRHD: DMAR Hardware Unit Definition Structure.

DmarEntryRecord
Represents the “header” of each dynamic table entry in the [DmarReporting] table.

DmarIter
An Iterator over the dynamic entries of the Dmar. Its lifetime is dependent upon the lifetime of its Dmar instance, which itself is bound to the lifetime of the underlying AcpiTables.

DmarRhsa
RHSA: DMAR Remapping Hardware Static Affinity Structure.

DmarRmrr
RMRR: DMAR Reserved Memory Region Reporting Structure.

DmarSatc
SATC: DMAR SoC Integrated Address Translation Cache Reorting Structure.

DrhdIter
An Iterator over the dynamic entries (DmarDeviceScopes) of the DmarDrhd. Its lifetime is dependent upon the lifetime of its DmarDrhd instance, which itself is bound to the lifetime of the underlying AcpiTables.

Enums

DmarEntry
The set of possible sub-tables that can exist in the top-level DMAR table.

Constants

DMAR_SIGNATURE

Functions

handle
The handler for parsing the DMAR table and adding it to the ACPI tables list.